fix : 공통 클래스 기준 장바구니 기능 수정
