+++ src/main/java/backend/mossy/boundedContext/market/app/product/MarketRegisterProductUseCase.java
@@ -1,4 +1,4 @@
-package backend.mossy.boundedContext.market.app;
+package backend.mossy.boundedContext.market.app.product;
 
 import backend.mossy.boundedContext.market.domain.Category;
 import backend.mossy.boundedContext.market.domain.MarketSeller;
